digest: let #finish raise if the output size is unknown
In OpenSSL 3.4 or later, SHAKE digests do not have a default output
length and EVP_MD_CTX_size() returns -1. Raise
OpenSSL::Digest::DigestError if the function returns a negative number.

Fortunately, this is currently caught by a sanity check in
rb_str_new() and leads to ArgumentError with the message "negative
string size (or size too big)". Still, the caller should check it.

Also, override #inspect to rescue an exception raised in #hexdigest.