digest: raise if the output size is unknown
The output length is unknown for XOFs such as SHAKE128 by default.

EVP_MD_CTX_size() returns -1 if the output length is unknown. This does
not lead to a crash thanks to the sanity check in rb_str_new(), which
raises ArgumentError with the message "negative string size (or size
too big)", but this should have been checked in ruby/openssl.