digest: add OpenSSL::Digest#squeeze
Add support for extendable-output functions (XOFs). An XOF is a hash
function that can produce arbitrary length of output. The interface of
the existing Digest::Class#digest and its variants is unsuitable for
this purpose.

Add a new method to "squeeze" the specified amount of output from the
XOF. This method can be called repeatedly. OpenSSL 3.3 or later and
AWS-LC support this.