asn1: decode BitString directly
Avoid accessing fields of ASN1_STRING, which is likely to become opaque
in OpenSSL 4.0.0.

We only use d2i_ASN1_BIT_STRING() for validating the first octet of the
contents octets. It can be trivially done ourselves.

This saves the allocation of ASN1_BIT_STRING object and one memory copy
of the contents. This also fixes a potential memory leak when
rb_str_new() raises an exception.