ssl: defer closing underlying TCP socket in start_server
Defer closing the underlying TCP socket until the end of start_server.
Depending on the platform and the OpenSSL version, doing so after an
unsuccessful SSLSocket#accept can close the TCP connection by RST. This
would not be a problem in a real scenario, but this prevents the client
from inspecting the TLS alert, which is useful for testing some code
paths in ruby/openssl.

Expand assert_handshake_error in test_ssl.rb as the only possible
exception now is OpenSSL::SSL::SSLError.

Add a dedicated test case for the rb_sys_fail() path in
SSLSocket#connect.